Changed an int to a size_t within PhaseList.h, and then took out warnings.

+++ Electrode/src/Electrode_CSTR.cpp
@@ -1949,7 +1949,7 @@ REDO:
             printf("\n");
         }
         printf("\t\t      SpeciesName        Moles_Init    Moles_final   SpMF       |   Src_Moles  Pred_Moles_Final\n");
-        for (int k = 0; k < m_NumTotSpecies; k++) {
+        for (size_t k = 0; k < m_NumTotSpecies; k++) {
             string ss = speciesName(k);
             double src =  DspMoles_final_[k] * deltaTsubcycleCalc_;
             printf("\t\t %20s  %12.4e  %12.4e  %12.4e | %12.4e %12.4e", ss.c_str(), spMoles_init_[k], spMoles_final_[k],
@@ -1961,7 +1961,7 @@ REDO:
                     int iStart = getGlobalSpeciesIndex(iph, 0);
                     for (int sp = 0; sp < numSpecInSolidPhases_[ph]; sp++) {
                         if (sp != phaseMFBig_[iph]) {
-                            int i = iStart + sp;
+                            size_t i = iStart + sp;
                             if (i == k) {
                                 found = true;
                                 double res =  spMoles_final_[k] - (spMoles_init_[k] + src);
@@ -2430,7 +2430,7 @@ void Electrode_CSTR::printElectrodePhase(int iphI, int pSrc, bool subTimeStep)
             }
             printf("\n");
             printf("                           spName                  SourceRateLastStep (kmol/m2/s) \n");
-            for (int k = 0; k <  m_NumTotSpecies; k++) {
+            for (size_t k = 0; k < m_NumTotSpecies; k++) {
                 string ss = speciesName(k);
                 printf("                           %-22s %10.3E\n", ss.c_str(), spNetProdPerArea[k]);
             }
@@ -2680,13 +2680,13 @@ void Electrode_CSTR::calcSrcTermsOnCompletedStep()
          *       An alternative would be to redo the residual calculation. However, here we assume that
          *       the residual calculation has been done and the results are in _final_
          */
-        for (int i = 0; i < m_NumTotSpecies; i++) {
+        for (size_t i = 0; i < m_NumTotSpecies; i++) {
             spMoleIntegratedSourceTermLast_[i] = spMoles_final_[i] - spMoles_init_[i];
         }
     } else {
         extractInfo();
         updateSpeciesMoleChangeFinal();
-        for (int isp = 0; isp < m_NumTotSpecies; isp++) {
+        for (size_t isp = 0; isp < m_NumTotSpecies; isp++) {
             spMoleIntegratedSourceTermLast_[isp] = DspMoles_final_[isp] * deltaTsubcycleCalc_;
         }
     }
@@ -2712,7 +2712,7 @@ void Electrode_CSTR::gatherIntegratedSrcPrediction()
     extractInfo();
     updateSpeciesMoleChangeFinal();
     IntegratedSrc_Predicted.resize(m_NumTotSpecies);
-    for (int isp = 0; isp < m_NumTotSpecies; isp++) {
+    for (size_t isp = 0; isp < m_NumTotSpecies; isp++) {
         IntegratedSrc_Predicted[isp] = DspMoles_final_[isp] * deltaTsubcycleCalc_;
     }
 }
